requirements.txt
python-binance
pandas
numpy<2
pandas-ta
requests
statsmodels
transformers
torch
scikit-learn
python-dotenv
